pyproject.toml
[project]
name = "schwab2pp"
description = "A command-line tool that converts a Charles Schwab transaction CSV file to a ready-to-import CSV file for Portfolio Performance."
authors = [{ name = "Rick Lan"}]
license = "MIT"
readme = "README.md"
# uv supports python 3.8 and higher
requires-python = ">=3.8,<4.0"
version = "1.0.1"
# click 8.2.0 errors on:
# no arguments: click.exceptions.MissingParameter: Missing parameter: name
# --help: TypeError: TyperArgument.make_metavar() takes 1 positional argument but 2 were given
dependencies = [
    "pandas >= 1.3.5",
    "typer-slim[standard]>=0.15.3",
    "click<8.2",
]

[dependency-groups]
dev = [
]
lint = [
    "codespell>=2.4.1",
    "ruff>=0.11.9",
]

[project.scripts]
schwab2pp = "schwab2pp.cli:app"

[project.urls]
Homepage = "https://github.com/rlan/convert-csv-schwab2pp"
Repository = "https://github.com/rlan/convert-csv-schwab2pp"
Documentation = "https://github.com/rlan/convert-csv-schwab2pp"
"Bug Tracker" = "https://github.com/rlan/convert-csv-schwab2pp/issues"

# ---- Build system ----

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


# ---- Settings ----

[tool.codespell]
# Ref: https://github.com/jlevy/simple-modern-uv/blob/06b288a9c1d032c7d6d0e4ed88c8515397529434/template/pyproject.toml.jinja
# Add here as needed:
# ignore-words-list = "foo,bar"
# skip = "foo.py,bar.py"

[tool.ruff]
# Ref: https://github.com/jlevy/simple-modern-uv/blob/06b288a9c1d032c7d6d0e4ed88c8515397529434/template/pyproject.toml.jinja
# Set as desired, typically 88 (black standard) or 100 (wide).
line-length = 88

[tool.ruff.lint]
# Ref: https://github.com/jlevy/simple-modern-uv/blob/06b288a9c1d032c7d6d0e4ed88c8515397529434/template/pyproject.toml.jinja
select = [
    # See: https://docs.astral.sh/ruff/rules/
    # Basic list from: https://docs.astral.sh/ruff/linter/#rule-selection
    "E", # https://docs.astral.sh/ruff/rules/#error-e
    "F", # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "UP", # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "B", # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "I", # https://docs.astral.sh/ruff/rules/#isort-i
    # Other possibilities:
    # "D" # https://docs.astral.sh/ruff/rules/#pydocstyle-d
    # "Q" # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    # "COM" # https://docs.astral.sh/ruff/rules/#flake8-commas-com
    # "SIM", # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim

]
ignore = [
    # Disable some rules that are overly pedantic. Add/remove as desired:
    "E501", # https://docs.astral.sh/ruff/rules/line-too-long/
    "E402", # https://docs.astral.sh/ruff/rules/module-import-not-at-top-of-file/
    "E731", # https://docs.astral.sh/ruff/rules/lambda-assignment/
    # We use both ruff formatter and linter so some rules should always be disabled.
    # See: https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191", # https://docs.astral.sh/ruff/rules/tab-indentation/
    "E111", # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple/
    "E114", # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple-comment/
    "E117", # https://docs.astral.sh/ruff/rules/over-indented/
    "D206", # https://docs.astral.sh/ruff/rules/docstring-tab-indentation/
    "D300", # https://docs.astral.sh/ruff/rules/triple-single-quotes/
    "Q000", # https://docs.astral.sh/ruff/rules/bad-quotes-inline-string/
    "Q001", # https://docs.astral.sh/ruff/rules/bad-quotes-multiline-string/
    "Q002", # https://docs.astral.sh/ruff/rules/bad-quotes-docstring/
    "Q003", # https://docs.astral.sh/ruff/rules/avoidable-escaped-quote/
    "COM812", # https://docs.astral.sh/ruff/rules/missing-trailing-comma/
    "COM819", # https://docs.astral.sh/ruff/rules/prohibited-trailing-comma/
    "ISC002", # https://docs.astral.sh/ruff/rules/multi-line-implicit-string-concatenation/
]

[tool.ruff.format]
docstring-code-format = true
